refactor(ast): split `Kroha.Syntax` into modules
