+++ src/Kroha/Backends/Nasm.hs
@@ -5,7 +5,7 @@ import Data.List (intercalate)
 import Data.Bifunctor (first)
 import Data.Maybe (fromJust)
 
-import Kroha.Syntax
+import Kroha.Syntax.Syntax
 import Kroha.Backends.Common
 import Kroha.Types
 import Kroha.Instructions (Instruction(..), LabelTarget(..), Target(..), Section)
