+++ src/Kroha/Stack.hs
@@ -4,7 +4,7 @@ import Control.Comonad (extract, duplicate, ($>))
 import Data.List (mapAccumL)
 import Data.Maybe (fromJust)
 
-import Kroha.Syntax
+import Kroha.Syntax.Syntax
 import Kroha.Types
 
 type StackRange = (Int, Int) {-offset, size-}
