+++ src/Kroha/Syntax/Syntax.hs
@@ -1,0 +1,9 @@
+module Kroha.Syntax.Syntax
+( module Kroha.Syntax.Primitive
+, module Kroha.Syntax.Statements
+, module Kroha.Syntax.Declarations
+) where
+
+import Kroha.Syntax.Primitive
+import Kroha.Syntax.Statements
+import Kroha.Syntax.Declarations
